vape-product-tagger/requirements.txt
# Core dependencies
pandas>=2.2.0
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Logging
colorlog>=6.8.0

# AI integration
ollama>=0.2.0

# Visualization (optional - for analytics)
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0

# Typing support (for Python < 3.9)
typing-extensions>=4.8.0

# HF Hub integration (for pulling fine-tuned models)
huggingface_hub>=0.19.0

# Training dependencies (install separately on Vast.ai or with: pip install -r vastai/requirements-train.txt)
# torch>=2.1.0
# transformers>=4.36.0
# peft>=0.7.0
# trl>=0.7.0
# bitsandbytes>=0.41.0
# accelerate>=0.25.0
# datasets>=2.15.0
# scikit-learn>=1.3.0
